+++ farmer/ncc/encoder_layer/mobilenetv2.py
@@ -1,0 +1,144 @@
+import tensorflow as tf
+
+from tensorflow.python.keras import layers
+from tensorflow.python.keras.layers import Input
+from tensorflow.python.keras.layers import Activation
+from tensorflow.python.keras.layers import Add
+from tensorflow.python.keras.layers import BatchNormalization
+from tensorflow.python.keras.layers import Conv2D
+from tensorflow.python.keras.layers import DepthwiseConv2D
+from tensorflow.python.keras.activations import relu
+
+
+def relu6(x):
+    return relu(x, max_value=6)
+
+
+def _make_divisible(v, divisor, min_value=None):
+    if min_value is None:
+        min_value = divisor
+    new_v = max(min_value, int(v + divisor / 2) // divisor * divisor)
+    # Make sure that round down does not go down by more than 10%.
+    if new_v < 0.9 * v:
+        new_v += divisor
+    return new_v
+
+
+def _inverted_res_block(inputs, expansion, stride, alpha, filters, block_id, skip_connection, rate=1):
+    in_channels = inputs.shape[-1]  # inputs._keras_shape[-1]
+    pointwise_conv_filters = int(filters * alpha)
+    pointwise_filters = _make_divisible(pointwise_conv_filters, 8)
+    x = inputs
+    prefix = 'expanded_conv_{}_'.format(block_id)
+    if block_id:
+        # Expand
+
+        x = Conv2D(expansion * in_channels, kernel_size=1, padding='same',
+                   use_bias=False, activation=None,
+                   name=prefix + 'expand')(x)
+        x = BatchNormalization(epsilon=1e-3, momentum=0.999,
+                               name=prefix + 'expand_BN')(x)
+        x = Activation(relu6, name=prefix + 'expand_relu')(x)
+    else:
+        prefix = 'expanded_conv_'
+    # Depthwise
+    x = DepthwiseConv2D(kernel_size=3, strides=stride, activation=None,
+                        use_bias=False, padding='same', dilation_rate=(rate, rate),
+                        name=prefix + 'depthwise')(x)
+    x = BatchNormalization(epsilon=1e-3, momentum=0.999,
+                           name=prefix + 'depthwise_BN')(x)
+
+    x = Activation(relu6, name=prefix + 'depthwise_relu')(x)
+
+    # Project
+    x = Conv2D(pointwise_filters,
+               kernel_size=1, padding='same', use_bias=False, activation=None,
+               name=prefix + 'project')(x)
+    x = BatchNormalization(epsilon=1e-3, momentum=0.999,
+                           name=prefix + 'project_BN')(x)
+
+    if skip_connection:
+        return Add(name=prefix + 'add')([inputs, x])
+
+    # if in_channels == pointwise_filters and stride == 1:
+    #    return Add(name='res_connect_' + str(block_id))([inputs, x])
+
+    return x
+
+
+def MobileNetV2(img_input, OS=16, alpha=1.):
+    """ Instantiates the Deeplabv3+ architecture
+
+    Optionally loads weights pre-trained
+    on PASCAL VOC or Cityscapes. This model is available for TensorFlow only.
+    # Arguments
+        OS: determines input_shape/feature_extractor_output ratio. One of {8,16}.
+            Used only for xception backbone.
+        alpha: controls the width of the MobileNetV2 network. This is known as the
+            width multiplier in the MobileNetV2 paper.
+                - If `alpha` < 1.0, proportionally decreases the number
+                    of filters in each layer.
+                - If `alpha` > 1.0, proportionally increases the number
+                    of filters in each layer.
+                - If `alpha` = 1, default number of filters from the paper
+                    are used at each layer.
+            Used only for mobilenetv2 backbone. Pretrained is only available for alpha=1.
+
+    # Returns
+        A Keras model instance.
+
+    """
+
+    OS = 8
+    first_block_filters = _make_divisible(32 * alpha, 8)
+    x = Conv2D(first_block_filters,
+                kernel_size=3,
+                strides=(2, 2), padding='same',
+                use_bias=False, name='Conv')(img_input)
+    x = BatchNormalization(
+        epsilon=1e-3, momentum=0.999, name='Conv_BN')(x)
+    x = Activation(relu6, name='Conv_Relu6')(x)
+
+    x = _inverted_res_block(x, filters=16, alpha=alpha, stride=1,
+                            expansion=1, block_id=0, skip_connection=False)
+
+    x = _inverted_res_block(x, filters=24, alpha=alpha, stride=2,
+                            expansion=6, block_id=1, skip_connection=False)
+    x = _inverted_res_block(x, filters=24, alpha=alpha, stride=1,
+                            expansion=6, block_id=2, skip_connection=True)
+
+    x = _inverted_res_block(x, filters=32, alpha=alpha, stride=2,
+                            expansion=6, block_id=3, skip_connection=False)
+    x = _inverted_res_block(x, filters=32, alpha=alpha, stride=1,
+                            expansion=6, block_id=4, skip_connection=True)
+    x = _inverted_res_block(x, filters=32, alpha=alpha, stride=1,
+                            expansion=6, block_id=5, skip_connection=True)
+
+    # stride in block 6 changed from 2 -> 1, so we need to use rate = 2
+    x = _inverted_res_block(x, filters=64, alpha=alpha, stride=1,  # 1!
+                            expansion=6, block_id=6, skip_connection=False)
+    x = _inverted_res_block(x, filters=64, alpha=alpha, stride=1, rate=2,
+                            expansion=6, block_id=7, skip_connection=True)
+    x = _inverted_res_block(x, filters=64, alpha=alpha, stride=1, rate=2,
+                            expansion=6, block_id=8, skip_connection=True)
+    x = _inverted_res_block(x, filters=64, alpha=alpha, stride=1, rate=2,
+                            expansion=6, block_id=9, skip_connection=True)
+
+    x = _inverted_res_block(x, filters=96, alpha=alpha, stride=1, rate=2,
+                            expansion=6, block_id=10, skip_connection=False)
+    x = _inverted_res_block(x, filters=96, alpha=alpha, stride=1, rate=2,
+                            expansion=6, block_id=11, skip_connection=True)
+    x = _inverted_res_block(x, filters=96, alpha=alpha, stride=1, rate=2,
+                            expansion=6, block_id=12, skip_connection=True)
+
+    x = _inverted_res_block(x, filters=160, alpha=alpha, stride=1, rate=2,  # 1!
+                            expansion=6, block_id=13, skip_connection=False)
+    x = _inverted_res_block(x, filters=160, alpha=alpha, stride=1, rate=4,
+                            expansion=6, block_id=14, skip_connection=True)
+    x = _inverted_res_block(x, filters=160, alpha=alpha, stride=1, rate=4,
+                            expansion=6, block_id=15, skip_connection=True)
+
+    x = _inverted_res_block(x, filters=320, alpha=alpha, stride=1, rate=4,
+                            expansion=6, block_id=16, skip_connection=False)
+    return x
+
